requirements.txt
fastapi
uvicorn
sqlalchemy
pillow
python-dotenv
